more code clean up
tests pass locally...now for GitHub...
