+++ jrnl/plugins/exporter/fancy.py
@@ -3,10 +3,11 @@
 # Copyright (C) 2012-2021 jrnl contributors
 # License: https://www.gnu.org/licenses/gpl-3.0.html
 
-from jrnl.exception import JrnlError
+
 from textwrap import TextWrapper
 
 from jrnl.plugins.base import BaseExporter
+from jrnl.plugins.util import check_provided_linewrap_viability
 
 from ... import __version__
 
@@ -83,14 +84,3 @@ class Exporter(BaseExporter):
     def export_journal(cls, journal):
         """Returns a unicode representation of an entire journal."""
         return "\n".join(cls.export_entry(entry) for entry in journal)
-
-
-def check_provided_linewrap_viability(linewrap, card, journal):
-    if len(card[0]) > linewrap:
-        width_violation = len(card[0]) - linewrap
-        raise JrnlError(
-            "LineWrapTooSmallForDateFormat",
-            config_linewrap=linewrap,
-            columns=width_violation,
-            journal=journal,
-        )
