+++ tests/test_export.py
@@ -1,7 +1,7 @@
-from jrnl.exception import JrnlError
-from jrnl.plugins.fancy_exporter import check_provided_linewrap_viability
-
 import pytest
+
+from jrnl.exception import JrnlError
+from jrnl.plugins.util import check_provided_linewrap_viability
 
 
 @pytest.fixture()
